refactor: 予約移動処理(dbup/dbdown/warikomi)を function_requestops.php へ切り出し
/api/request_move.php と delete.php で共用するため関数本体を無改修で
移動。delete.php の挙動は不変。
